Add package.{Describe,List,Delete}
This change adds the remainder of the ServiceRunner primitives
describe, list, and delete. The autogenerated code will be included in
a subsequent commit in order to make sure packages importing the
golang sdk will continue to work. The autogenerated code is being
committed separately to avoid polluting the commit diff with noise.