Add optional Service.ExecAssets property
Executables produced by non-golang programming languages (e.g. Java)
may require library or other resources to be included along with the
executable binary in order to run correctly. This change introduces an
ExecAssets property to a service to allow a Bopmatic project to define
the set of files which should be included in a package submission. The
Executable property is now relative to the ExecAssets directory when
specified.

Additionally, this change also removes the ImportPath property of a
Bopmatic project as this is unused and not needed.