Add support for ApiDefAssets
More complex GRPC .proto service definitions may include nested
imports. In this situation these nested imports also need to be
included in the Bopmatic package. This change introduces a new
Bopmatic.yaml service field, apidef_assets, which works similarly to
exec_assets. Customers can optionally define a directory of .proto
files to be included in the package. The testcase oneSvcWithApiAssets
shows an example of how this can be done.