Update auto-generated code
The prior dependencies update results in diffs in the autogenerated
code. This commit updates these autogenerated source files. We include
these in the git repository in to enable downstream projects which
depend on the Bopmatic sdk to compile correctly without needing to
take dependencies on the toolchain we use to autogenerate them.