SR API refactor Step 1 (part 3 of 3)
This commit is part of a series which undertakes a major refactor
of SR's public API. The series will introduce many new primitives in
the public facing API including:

1. Project primitives: CreateProject, DeleteProject, ListProjects, and
  DescribeProject
2. Environment primitives: CreateEnvironment, DeleteEnvironment,
3. ListEnvironments, DescribeEnvironment
4. Project detail primitives: DescribeSite, ListServices,
  DescribeService, ListDatastores, DescribeDatastore, ListDatabases,
  DescribeDatabase
5. Deployment detail primitives: CreateDeployment, ListDeployments,
  DescribeDeployment
6. Package primitives: UploadPackage
7. Metric primitives: ListMetrics, GetMetricSamples

The principle motivation for the introduction of these primitives is
to support the development of Bopmatic's UI console. Conceptual
changes compared to preview iterations of the API are discussed below.

1. Environments. Previously SR only supported a single implicit prod
environment. Now projects can be deployed into prod, dev&test,
individual developer, or an arbitrary number of environments that
customers wish to create.

2. Deployments. Previously SR coupled the notions of pacakge with
deployments. However, in order to make infrastructure updates that
Bopmatic performs on behalf of customers, and also allow the same
package version to be deployed into multiple environments, we
introduce deployments as a distinct primitive.

Note that this is a breaking change. Consumers of the SDK
(e.g. Bopmatic's CLI) will have to refactor when updating.

Step 1 of the series refactors the existing API implementations.

Step 1.Part 3 contains only the changes to the autogenerated code