Extend {Database,Service}Description w/ additional fields (part 1 of 2)
This commit extends ServiceDescription with additional information
including the set of databases and datastores the service is
configured to access. It also adds the api port and api definition
that the service is configured with.

This commit also extends DatabaseDescription to include more per-table
information rather than just a list of table names.